Unify HTML, markdown, and file preview into one sheet
AttachmentPreviewSheet replaces the per-type HTMLAttachmentPreviewSheet,
the inline private MarkdownAttachmentPreviewSheet in MessagesViewController,
and the FileAttachmentQuickLookCoordinator path with a single SwiftUI sheet
that branches on attachment.isHTMLFile / .isMarkdownFile / else.

NavigationStack toolbar holds the standard iOS 26 cancellation/principal/
confirmation items so the close X, sender pill, and share button render as
the same floating glass capsules ConversationView uses (matching pill style,
interactive dismiss working out of the box, taps actually firing).

Sender pill is a new AttachmentSenderIndicator showing the avatar, display
name, and a relative date (Today / Yesterday / weekday / dated string).

Content extends under the toolbar via .ignoresSafeArea() and the toolbar bar
background is hidden so the floating pills naturally blur whatever's behind
them, identical to ConversationView. To make the area behind the pills feel
seamless when there's nothing to blur yet, AttachmentHTMLContent posts the
body's computed background color via a WKScriptMessageHandler and the sheet
applies it as the content's .background(...), so the pills pick up the
artifact's actual tint via their glass effect.

Generic files render via QLPreviewController embedded in a UINavigationController
with its own nav bar hidden, so QuickLook keeps handling every file format
(PDF, image, video, audio, etc.) while our toolbar floats on top.

Threaded the message context through the onOpenFile chain (now
(HydratedAttachment, AnyMessage) -> Void) across the data source, cell
factory, group view, group item view, and reply reference view, so the
sheet can render the sender's profile + the message's sentAt date in the
pill. Reply parents wrap their parentMessage as .message(parentMessage,
.existing) so taps open the parent's attachment with the parent's sender
and date in the pill.

Deleted HTMLAttachmentPreviewSheet.swift and the inline
MarkdownAttachmentPreviewSheet/MarkdownWebView from MessagesViewController.
FileAttachmentQuickLookCoordinator stays for AssistantFilesLinksView's
separate file-browser flow.
