+++ Convos/Conversation Detail/Messages/MessagesListView/MessagesGroupItemView.swift
@@ -15,7 +15,7 @@ struct MessagesGroupItemView: View {
     let onPhotoRevealed: (String) -> Void
     let onPhotoHidden: (String) -> Void
     let onPhotoDimensionsLoaded: (String, Int, Int) -> Void
-    var onOpenFile: ((HydratedAttachment) -> Void)?
+    var onOpenFile: ((HydratedAttachment, AnyMessage) -> Void)?
     var onTapReactions: ((AnyMessage) -> Void)?
     var onReaction: ((String, String) -> Void)?
     let onToggleReaction: (String, String) -> Void
@@ -301,7 +301,7 @@ struct MessagesGroupItemView: View {
             )
             .id(message.messageId)
         } else if attachment.isHTMLFile {
-            let htmlTapAction: () -> Void = { onOpenFile?(attachment) }
+            let htmlTapAction: () -> Void = { onOpenFile?(attachment, message) }
             let avatarTap: () -> Void = { onTapAvatar(message) }
             let reactionsTap: () -> Void = { onTapReactions?(message) }
             HTMLAttachmentBubble(
@@ -320,7 +320,7 @@ struct MessagesGroupItemView: View {
             )
             .id(message.messageId)
         } else if attachment.mediaType == .file {
-            let fileTapAction: () -> Void = { onOpenFile?(attachment) }
+            let fileTapAction: () -> Void = { onOpenFile?(attachment, message) }
             FileAttachmentBubble(
                 attachment: attachment,
                 style: bubbleType,
